Replace `FeedPresenter` from iOS module with `EssentialFeed`
The presenter is cross-platform, therefore it should not
live in the iOS module, but rather in the Essential Feed
module
