+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedErrorViewModel.swift
@@ -1,5 +1,5 @@
-struct FeedErrorViewModel {
-    let message: String?
+public struct FeedErrorViewModel {
+    public let message: String?
     
     static var noError: Self {
         FeedErrorViewModel(message: nil)
