+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedLoadingViewModel.swift
@@ -1,0 +1,3 @@
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
