+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedViewModel.swift
@@ -1,0 +1,3 @@
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
