--- EssentialFeed/EssentialFeediOS/Feed Presentation/FeedLoadingViewModel.swift
@@ -1,3 +1,0 @@
-struct FeedLoadingViewModel {
-    let isLoading: Bool
-}
