--- EssentialFeed/EssentialFeediOS/Feed Presentation/FeedViewModel.swift
@@ -1,5 +1,0 @@
-import EssentialFeed
-
-struct FeedViewModel {
-    let feed: [FeedImage]
-}
