--- EssentialFeed/EssentialFeediOS/Feed Presentation/en.lproj/Feed.strings
@@ -1,2 +1,0 @@
-"FEED_TITLE_KEY" = "My Feed";
-"FEED_VIEW_CONNECTION_ERROR" = "Couldn't connect to server";
