--- EssentialFeed/EssentialFeediOS/Feed Presentation/sv.lproj/Feed.strings
@@ -1,2 +1,0 @@
-"FEED_TITLE_KEY" = "Mitt flöde";
-"FEED_VIEW_CONNECTION_ERROR" = "Kunde inte koppla till servern";
