+++ EssentialFeed/EssentialFeediOSTests/Feed Presentation/FeedLocalizationTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import EssentialFeediOS
+import EssentialFeed
 
 final class FeedLocalizationTests: XCTestCase {
