+++ EssentialFeed/EssentialFeediOSTests/FeedUI/Helpers/FeedViewController+Localization.swift
@@ -1,11 +1,11 @@
 import Foundation
-import EssentialFeediOS
+import EssentialFeed
 import XCTest
 
 extension FeedUIIntegrationTests {
     func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
         let table = "Feed"
-        let bundle = Bundle(for: FeedViewController.self)
+        let bundle = Bundle(for: FeedPresenter.self)
         let value = bundle.localizedString(forKey: key, value: nil, table: table)
         
         if value == key {
